% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_temperatura_mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{Grafico de temperatura media mensual por estacion}
\usage{
grafico_temperatura_mensual(datos, colores = NULL, titulo = "Temperatura")
}
\arguments{
\item{datos}{Tibble con columnas: id, fecha (Date), temperatura_abrigo_150cm.}

\item{colores}{Vector de colores opcional.}

\item{titulo}{Titulo del grafico.}
}
\value{
Un objeto ggplot2.
}
\description{
Calcula la temperatura media mensual por estacion y genera un grafico
usando ggplot2.
}
\examples{
# Cargar datos internos
data("NH0472")

# Graficar la temperatura mensual
grafico_temperatura_mensual(NH0472)

}
