% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_estacion.R
\name{leer_estacion}
\alias{leer_estacion}
\title{Lee y descarga datos meteorologicos de una estacion}
\usage{
leer_estacion(id_estacion, ruta)
}
\arguments{
\item{id_estacion}{Cadena de texto con el identificador de la estacion.}

\item{ruta}{Cadena de texto indicando donde guardar o leer el archivo CSV.}
}
\value{
Un tibble con los datos meteorologicos de la estacion.
}
\description{
Descarga (si es necesario) y lee un archivo CSV con datos
meteorologicos correspondientes a una estacion especifica.
}
\examples{
data("NH0472")
head(NH0472)

}
