% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NH0472.R
\docType{data}
\name{NH0472}
\alias{NH0472}
\title{Dataset NH0472}
\format{
Un data frame con varias columnas numericas y de fecha.
}
\source{
Estacion meteorologica NH0472
}
\usage{
NH0472
}
\description{
Datos meteorologicos de la estacion NH0472 incluidos en el paquete
para ejemplos, tests y demostraciones.
}
\examples{
data("NH0472")
head(NH0472)
}
\keyword{datasets}
