% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NH0472.R
\docType{data}
\name{NH0472}
\alias{NH0472}
\title{Datos de la estacion NH0472}
\format{
Un archivo CSV con columnas tipicas de estaciones meteorologicas.
}
\source{
Estacion meteorologica NH0472.
}
\usage{
NH0472
}
\description{
Conjunto de datos meteorologicos reales de la estacion "NH0472".
Este archivo se incluye en \verb{inst/extdata/} y se utiliza para ejemplos,
tests y demostraciones del paquete.
}
\details{
Para acceder a este archivo use:

\code{system.file("extdata", "NH0472.csv", package = "meteopisa")}
}
\examples{
archivo <- system.file("extdata", "NH0472.csv", package = "meteopisa")
datos <- readr::read_csv(archivo, show_col_types = FALSE)
head(datos)
}
\keyword{datasets}
