% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_resumen_temperatura.R
\name{tabla_resumen_temperatura}
\alias{tabla_resumen_temperatura}
\title{Resumen estadistico de temperatura por estacion}
\usage{
tabla_resumen_temperatura(ids_estaciones = NULL, datos_lista = NULL)
}
\arguments{
\item{ids_estaciones}{Vector opcional de IDs de estaciones. Se usa si
\code{datos_lista} es NULL.}

\item{datos_lista}{Lista opcional de data frames ya cargados, con nombres
correspondientes a los IDs de estacion.}
}
\value{
Un tibble en formato largo con columnas:
\code{id}, \code{Tipo}, \code{Temperatura}.
}
\description{
Calcula media, desvio estandar, maxima y minima de temperatura para una o mas
estaciones. Puede funcionar en dos modos:
\enumerate{
\item Usando datasets internos (\code{datos_lista}), ideal para examples y tests.
\item Descargando o leyendo archivos reales mediante \code{leer_estacion()}.
}
}
\examples{
# Usando el dataset interno (modo recomendado para examples)
data("NH0472")

tabla_resumen_temperatura(
  datos_lista = list(estacion_NH0472 = NH0472)
)

}
